Delete README.md
File Structure:
I have created an abstract class called Firewall and implemented its methods in Accepting_packets.java file, I did this so that It will be easy to add more functionalities(like discarding packets based on rules) to the Firewall class. 

Run-Time:
I have used HashSet which is having the search time complexity O(1). So the number of records in the file having all the rules will not affect the time taken to process the packet. Assuming that the frequency of the file(Rules) processing is less than processing incoming packets.

How to Run:
All the logic is implemented in the package Firewall, I have also uploaded a java file where this package functionalities are used.
1. Run package.java, 
2. You can enter the file or you can use a sample file(which I have uploaded).
3. You can enter the inputs in the following format
     a. Input should be of four words separated by spaces.
     b. Order of input should as follows, direction protocol port_address IP address
4. The program prompts the result.